requirements.txt
# CORE LIBRARIES
numpy==1.26.4
pandas==2.3.2
scipy==1.16.2
scikit-learn==1.7.2
statsmodels==0.14.5

# VISUALIZATION
matplotlib==3.10.6
seaborn==0.13.2

# DATE, TIME & UTILS
PyYAML==6.0.2

# PROJECT LIBRARIES
darts==0.30.0
mlflow==2.15.1
sphinx==7.4.7
sphinx-rtd-theme==2.0.0
